binding: fix excessively large channel count param types
both `Map` and `ChannelVolumes` hold a fixed size array, exactly the
same as the underlying C types, with the length of that array being
`CHANNELS_MAX`, which is currently set to 32. additionally, they both
include an attribute indicating how much of the arrays are in use,
which in both cases is a `u8`.

there were some functions using a `u32` to take a channel count, which
is silly since the count is limited to the range of a `u8` in those
cases. the use of `u32` simply came from the fact the underlying C
functions they call take an `int`.

there is no need for a type larger than `u8`. changing to `u8` better
reflects to users what sort of range of values are expected.

+++ pulse-binding/src/channelmap.rs
@@ -393,10 +393,10 @@ impl Map {
     ///
     /// This call will fail (return `None`) if there is no default channel map known for this
     /// specific number of channels and mapping.
-    pub fn init_auto(&mut self, channels: u32, def: MapDef) -> Option<&mut Self> {
-        debug_assert!(channels <= Self::CHANNELS_MAX as u32);
+    pub fn init_auto(&mut self, channels: u8, def: MapDef) -> Option<&mut Self> {
+        debug_assert!(channels <= Self::CHANNELS_MAX);
         unsafe {
-            if capi::pa_channel_map_init_auto(self.as_mut(), channels, def).is_null() {
+            if capi::pa_channel_map_init_auto(self.as_mut(), channels as u32, def).is_null() {
                 return None;
             }
         }
@@ -406,9 +406,9 @@ impl Map {
     /// Similar to [`init_auto`](#method.init_auto) but instead of failing if no default mapping is
     /// known with the specified parameters it will synthesize a mapping based on a known mapping
     /// with fewer channels and fill up the rest with AUX0...AUX31 channels.
-    pub fn init_extend(&mut self, channels: u32, def: MapDef) -> &mut Self {
-        debug_assert!(channels <= Self::CHANNELS_MAX as u32);
-        unsafe { capi::pa_channel_map_init_extend(self.as_mut(), channels, def) };
+    pub fn init_extend(&mut self, channels: u8, def: MapDef) -> &mut Self {
+        debug_assert!(channels <= Self::CHANNELS_MAX);
+        unsafe { capi::pa_channel_map_init_extend(self.as_mut(), channels as u32, def) };
         self
     }
